Move Build-Depends for documentation to Build-Depends-Indep
This reduces the list of build-dependencies on arch-any builds quite a lot
and should allow bind9 to build on s390x where xindy is currently not
available.

The change has been verified by building once with

sbuild --arch-all --arch-any

and once with

sbuild --arch-all --no-arch-any
sbuild --no-arch-all --arch-any

and comparing the result.